Move UI to info pane
