+++ src/App.vue
@@ -4,17 +4,17 @@
             <main-view v-if="!game.features.prestige.prestiged" :game="game"/>
             <prestige-view v-if="game.features.prestige.prestiged" :game="game"/>
         </div>
-        <plot-modal :show="game.features.controller.openedModal == ModalType.Plot" @close="closeModal"/>
-        <wiki-modal :show="game.features.controller.openedModal == ModalType.Wiki" @close="closeModal"/>
-        <settings-modal :show="game.features.controller.openedModal == ModalType.Settings" @close="closeModal"/>
-        <achievements-modal :show="game.features.controller.openedModal == ModalType.Achievements" @close="closeModal"/>
-        <prestige-modal :show="game.features.controller.openedModal == ModalType.Prestige" @close="closeModal"/>
+        <plot-modal :show="openedModal == InfoType.Plot" @close="closeModal"/>
+        <wiki-modal :show="openedModal == InfoType.Wiki" @close="closeModal"/>
+        <settings-modal :show="openedModal == InfoType.Settings" @close="closeModal"/>
+        <achievements-modal :show="openedModal == InfoType.Achievements" @close="closeModal"/>
+        <prestige-modal :show="openedModal == InfoType.Prestige" @close="closeModal"/>
     </div>
 </template>
 
 <script>
 import {App} from "@/App.ts"
-import {ModalType} from '@/scripts/GameController';
+import {InfoType} from '@/scripts/GameController';
 
 import PlotModal from '@/controls/farm/plot-modal/plot-modal.vue';
 import WikiModal from '@/controls/wiki/wiki-modal.vue';
@@ -37,18 +37,18 @@ export default {
     data() {
         return {
             game: App.game,
-            ModalType,
+            InfoType,
         }
     },
     computed: {
         darkMode() {
-            return App.game.features.settings.getSetting('darkMode').value;
+            return this.game.features.settings.getSetting('darkMode').value;
         },
+        openedModal() {
+            return this.game.features.controller.openedModal;
+        }
     },
     methods: {
-        openModal(modalType) {
-            this.game.features.controller.openModal(modalType);
-        },
         closeModal() {
             this.game.features.controller.closeModal();
         }
