+++ src/controls/main-view.vue
@@ -11,28 +11,28 @@
               <div class="iconBox">
                     <!-- Additional Icons -->
                     <div class="flex flex-wrap gap-1 justify-center">
-                        <div class="btn" style="height:32px;" @click="openModal(ModalType.Wiki)">
+                        <div class="btn" style="height:32px;" @click="openInfo(InfoType.Wiki)">
                             <icon class="has-tooltip" :bg="game.features.notifications.hasWikiNotification ? 'bg-red-500' : 'bg-icon'" :image="require(`@/assets/images/icons/Wiki Icon.png`)">
                                 <tooltip position="top-left" :interactable="false">
                                 <div class="text-center">Wiki</div>
                                 </tooltip>
                             </icon>
                         </div>
-                        <div class="btn" style="height:32px;" @click="openModal(ModalType.Achievements)">
+                        <div class="btn" style="height:32px;" @click="openInfo(InfoType.Achievements)">
                             <icon class="has-tooltip" :image="require(`@/assets/images/icons/Achievement Icon.png`)">
                                 <tooltip position="top-left" :interactable="false">
                                 <div class="text-center">Achievements</div>
                                 </tooltip>
                             </icon>
                         </div>
-                        <div class="btn" style="height:32px;" @click="openModal(ModalType.Prestige)">
+                        <div class="btn" style="height:32px;" @click="openInfo(InfoType.Prestige)">
                             <icon class="has-tooltip" :image="require(`@/assets/images/icons/Prestige Icon.png`)">
                                 <tooltip position="top-left" :interactable="false">
                                 <div class="text-center">Prestige</div>
                                 </tooltip>
                             </icon>
                         </div>
-                        <div class="btn" style="height:32px;" @click="openModal(ModalType.Settings)">
+                        <div class="btn" style="height:32px;" @click="openInfo(InfoType.Settings)">
                             <icon class="has-tooltip" :image="require(`@/assets/images/icons/Settings Icon.png`)">
                                 <tooltip position="top-left" :interactable="false">
                                 <div class="text-center">Settings</div>
@@ -72,7 +72,15 @@
         </div>
         <div id="right-column" class="col-span-10 lg:col-span-5">
             <div class="flex flex-col gap-1" style="height: 100%;">
-
+                <info-view style="height: 100%;" 
+                    :controller="controller" 
+                    :farms="farms" 
+                    :wiki="wiki" 
+                    :prestige="prestige"
+                    :settings="settings"
+                    :beans="beans"
+                    :plants="plants"
+                    :notifications="notifications" />
             </div>
         </div>
     </div>
@@ -81,7 +89,7 @@
 <script>
 import {App} from "@/App.ts"
 
-import {ModalType} from '@/scripts/GameController';
+import {InfoType} from '@/scripts/GameController';
 
 import ToolIcons from '@/controls/controller/tool-icons/tool-icons.vue';
 import BeanList from '@/controls/controller/beanlist/bean-list.vue';
@@ -92,6 +100,7 @@ import FarmControl from "@/controls/farm/farm-control.vue";
 import Tooltip from '@/controls/utility/tooltip.vue';
 import { Game } from '@/Game';
 import IgtDeveloperPanel from '@/components/developer-panel/igt-developer-panel.vue';
+import InfoView from './info-view.vue';
 
 export default {
     components: {
@@ -103,10 +112,11 @@ export default {
         FarmControl,
         Tooltip,
         IgtDeveloperPanel,
+        InfoView,
     },
     data() {
         return {
-            ModalType,
+            InfoType,
         }
     },
     props: {
@@ -131,15 +141,42 @@ export default {
                 'image-rendering': 'pixelated',
                 'height': '60vh',
             };
+        },
+        
+        //#region Features
+        settings() {
+            return this.game.features.settings;
+        },
+        controller() {
+            return this.game.features.controller;
+        },
+        upgrades() {
+            return this.game.features.upgrades;
+        },
+        beans() {
+            return this.game.features.beans;
+        },
+        plants() {
+            return this.game.features.plants;
+        },
+        farms() {
+            return this.game.features.farms;
+        },
+        prestige() {
+            return this.game.features.prestige;
+        },
+        wiki() {
+            return this.game.features.wiki;
+        },
+        notifications() {
+            return this.game.features.notifications;
         }
+        //#endregion
     },
     methods: {
-        openModal(modalType) {
-            this.game.features.controller.openModal(modalType);
+        openInfo(infoType) {
+            this.game.features.controller.openInfo(infoType);
         },
-        closeModal() {
-            this.game.features.controller.closeModal();
-        }
     },
 }
 </script>
